Fix test failures due to hidden dependency
The test for managed dependencies uses a map as a project definition
and attempts to run tests through doo. However, it uses a manged
dependency on org.clojure/clojure, the version for which may or may
not be the same as what's in the project.clj for doo.

Rather than try to keep them in sync, having a test profile which
brings in the tested version makes the dependency explicit. It doesn't
need to be Clojure 1.8.0 for any reason, it just needs to be
a *specific* version.