Add library to install and deploy tasks, but not script
Running lein test for the doo library seems to require installing a
bunch of dependencies in the CI system. I assume it's trying to run
tests in a bunch of the different environments.

The thoroughness is admirable, and I'd want those tests running if
we're going to maintain this long-term. But I hope we won't have to,
so I'm leaving it out for now.